requirements.txt
flask>=3.0
flask-socketio>=5.3
requests>=2.28
openai>=1.0
numpy>=1.24
scipy>=1.10
pyserial>=3.5
python-socketio>=5.9
eventlet>=0.33
